Move GCEGatewayService under GCENetwork
