+++ cloudbridge/cloud/providers/gce/services.py
@@ -8,7 +8,6 @@ from cloudbridge.cloud.base.resources import ClientPagedResultList
 from cloudbridge.cloud.base.resources import ServerPagedResultList
 from cloudbridge.cloud.base.services import BaseBucketService
 from cloudbridge.cloud.base.services import BaseComputeService
-from cloudbridge.cloud.base.services import BaseGatewayService
 from cloudbridge.cloud.base.services import BaseImageService
 from cloudbridge.cloud.base.services import BaseInstanceService
 from cloudbridge.cloud.base.services import BaseKeyPairService
@@ -33,7 +32,6 @@ from retrying import retry
 
 from .resources import GCEFirewallsDelegate
 from .resources import GCEInstance
-from .resources import GCEInternetGateway
 from .resources import GCEKeyPair
 from .resources import GCEMachineImage
 from .resources import GCENetwork
@@ -566,7 +564,6 @@ class GCENetworkingService(BaseNetworkingService):
         self._network_service = GCENetworkService(self.provider)
         self._subnet_service = GCESubnetService(self.provider)
         self._router_service = GCERouterService(self.provider)
-        self._gateway_service = GCEGatewayService(self.provider)
 
     @property
     def networks(self):
@@ -763,26 +760,6 @@ class GCERouterService(BaseRouterService):
         except googleapiclient.errors.HttpError as http_error:
             cb.log.warning('googleapiclient.errors.HttpError: %s', http_error)
             return None
-
-
-class GCEGatewayService(BaseGatewayService):
-    _DEFAULT_GATEWAY_NAME = 'default-internet-gateway'
-    _GATEWAY_URL_PREFIX = 'global/gateways/'
-
-    def __init__(self, provider):
-        super(GCEGatewayService, self).__init__(provider)
-        self._default_internet_gateway = GCEInternetGateway(
-            provider,
-            {'id': (GCEGatewayService._GATEWAY_URL_PREFIX +
-                    GCEGatewayService._DEFAULT_GATEWAY_NAME),
-             'name': GCEGatewayService._DEFAULT_GATEWAY_NAME})
-
-    def get_or_create_inet_gateway(self, name):
-        GCEInternetGateway.assert_valid_resource_name(name)
-        return self._default_internet_gateway
-
-    def delete(self, gateway):
-        pass
 
 
 class GCESubnetService(BaseSubnetService):
